fix: useInput 분리
